refactor(layout): extract gutter detection
Move recurring gutter and side-panel split detection into a dedicated layout module so line construction can reuse named boundary predicates.
